for debug

+++ sofa-jobs/sofa-jobs-impl/src/main/java/me/izhong/jobs/manage/impl/core/model/XxlJobInfo.java
@@ -36,16 +36,16 @@ public class XxlJobInfo  extends TimedBasedEntity implements Serializable {
 	private String executorParam;		    // 执行器，任务参数
 	private String executorBlockStrategy;	// 阻塞处理策略
 	private Long executorTimeout;     		// 任务执行超时时间，单位秒
-	private Integer executorFailRetryCount;		// 失败重试次数
+	private Long executorFailRetryCount;		// 失败重试次数
 	
-	private String glueType;		// GLUE类型	#com.xxl.job.core.glue.GlueTypeEnum
+	private String glueType;		// GLUE类型
 	private String glueSource;		// GLUE源代码
 	private String glueRemark;		// GLUE备注
 	private Date glueUpdatetime;	// GLUE更新时间
 
 	private String childJobId;		// 子任务ID，多个逗号分隔
 
-	private Integer triggerStatus;		// 调度状态：0-停止，1-运行
+	private Long triggerStatus;		// 调度状态：0-停止，1-运行
 	private Long triggerLastTime;	// 上次调度时间
 	private Long triggerNextTime;	// 下次调度时间
 }
